# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing coin calculations.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/mincoins": {
            "get": {
                "description": "Calculates the minimum number of coins needed for a given amount using the largest denominations first.",
                "summary": "Minimum number of coins",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Amount to change",
                        "name": "val",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Available denominations",
                        "name": "coins",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of coins that sum up to the amount",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        },
        "/mincoins2": {
            "get": {
                "description": "Calculates the minimum number of coins needed for a given amount, optimizes by removing duplicates and sorting denominations in descending order.",
                "summary": "Minimum number of coins with optimization",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Amount to change",
                        "name": "val",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Available denominations",
                        "name": "coins",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Optimized list of coins that sum up to the amount",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        }
    }
}